+++ Models/ProductInputModel.cs
@@ -14,7 +14,7 @@ namespace PROGETTO_SETTIMINALE_BE_S2_L5__Vescio_Pia_Francesca.Models
 
         [Display(Name = "Prezzo")]
         [DisplayFormat(DataFormatString = "{0:C}", ApplyFormatInEditMode = true)]
-        public decimal Price { get; set; } = 0;
+        public int Price { get; set; } = 0;
 
         [Display(Name = "Immagine principale")]
         [Required(ErrorMessage = "Il campo {0} è obbligatorio. Inseriscilo correttamente.")]
